[all] move tests to unit tests using real production types (#90)
